+++ qcloud_sdk/__init__.py
@@ -5,3 +5,9 @@ from qcloud_sdk.api import APIClient
 
 # 各产品APIClient
 from qcloud_sdk.cos.client import CosAPIClient
+
+# 异常类
+from qcloud_sdk.exceptions import QCloudAPIException
+
+# 声明式配置
+from qcloud_sdk.config import settings