+++ qcloud_sdk/sms/__init__.py
@@ -1,12 +1,6 @@
-# -*- coding: utf-8 -*-
 """
 腾讯云SMS短信服务SDK
 """
-
-# API
-from .api import QCloudSMSAPIClient
-# 异常类
-from .exception import QCloudSMSAPIException